# cache: some optimizations