@fturib 2018-10-29
`BenchmarkCacheResponse-4 9599 8735 -9.00%` : is that mean we have a 9% gain in perf on the cache.

That worth the change !