@miekg 2018-10-29
[ Quoting <notifications@github.com> in "Re: [coredns/coredns] cache: some o..." ]
>`BenchmarkCacheResponse-4 9599 8735 -9.00%` : is that mean we have a 9% gain in perf on the cache.

Going back to strings (qname + qtype + do), was 15%, but adds more memory 
pressure. Also need way more changes, but might be a things we need to go back 
to.